+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 2.9.9
 - Fix Delete Prompt displayed when working User/Student cleared program wide
 - Add Warning if not in current Quarter in Assignments.php
 - Check IDs are valid for current school & syear in MarkingPeriods.php & Assignments.php
+- Help texts updates in Help_en.php & Help_es.php
 
 Changes in 2.9.8
 ----------------